catalogue/pyproject.toml
[tool.poetry]
name = "ai-agent-service"
version = "0.1.0"
description = "A service for managing AI agents"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
python-multipart = "^0.0.6"
pydantic = "^2.4.2"
pyyaml = "^6.0.1"
boto3 = "^1.34.0"
pydantic-settings = "^2.0.0"
botocore = "^1.36.16"
[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 